requirements.txt
Django==4.2.1
Pillow==9.5.0
django-autoslug==1.9.9
Unidecode==1.3.6